--- modules/core/src/constants/index.js
@@ -1,8 +1,0 @@
-export {default as VIEW_MODES} from './view-modes';
-
-export const COORDINATES = {
-  GEOGRAPHIC: 'geographic',
-  MAP_RELATIVE: 'map_relative',
-  VEHICLE_RELATIVE: 'vehicle_relative',
-  CUSTOM: 'custom'
-};
